+++ packaging/chocolatey/tools/chocolateyinstall.ps1
@@ -1,9 +1,9 @@
 $ErrorActionPreference = 'Stop'
 $toolsDir = "$(Split-Path -Parent $MyInvocation.MyCommand.Definition)"
 
-$tag    = 'v1.3.3'
+$tag    = 'v1.3.4'
 $url    = "https://github.com/Cod-e-Codes/marchat/releases/download/$tag/marchat-$tag-windows-amd64.zip"
-$checksum = '6d8a2cceb142f8e428f2f4b01c980ff3b1e0121d2deb6f88d1f5f011be5527af'
+$checksum = '0000000000000000000000000000000000000000000000000000000000000000'
 
 $packageArgs = @{
   packageName   = $env:ChocolateyPackageName